sharedtex_mt: fix rendering thread hang
XNextEvent is a blocking call which locks up the display mutex.  This
causes the rendering threads to hang when they call glXSwapBuffers() as
that tries to take the same mutex in underlying calls through XCopyArea().
So we only call XNextEvent when it has at least one event so we don't lock
indefinitely.

Patch slightly fixed by Brian to keep var declarations as-is and clean up
the patch description.

Signed-off-by: Awais Belal <awais_belal@mentor.com>
Reviewed-by: Brian Paul <brianp@vmware.com>